+++ lib/home_page.dart
@@ -1,0 +1,47 @@
+import 'package:flutter/material.dart';
+import 'second_page.dart';
+
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({super.key});
+
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+
+  @override
+  Widget build(BuildContext context) {
+    void importCSV() => (
+
+    );
+
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        title: const Text("upload your csv data"),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            ElevatedButton.icon(         
+              onPressed: importCSV,              
+              icon: const Icon(Icons.file_upload),
+              label: const Text('Import CSV'),
+        )
+        ],
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: (){
+          Navigator.push(
+            context,
+            MaterialPageRoute(builder: (context) => SecondPage())
+            );
+        },
+        child: const Icon(Icons.last_page),
+        ),
+    );
+  }
+}
