+++ lib/second_page.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class SecondPage extends StatelessWidget{
+  @override
+  Widget build(BuildContext context){
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        title: const Text("Second Page"),
+      ),
+    );
+  }
+}
